# packages/server/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/devices/{id}/actions/tap": {
      "post": {
        "operationId": "CommandsController_tap",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TapDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AckCommandResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "DTO 검증 실패·미지원 액션",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "토큰 없음·불일치",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "점유자 불일치",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "기기 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "기기 오프라인",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "요청 과다 (rate limit)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "502": {
            "description": "Agent 미연결·기기 명령 실패",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "504": {
            "description": "기기 응답 시간 초과",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "좌표 탭",
        "tags": [
          "commands"
        ]
      }
    },
    "/devices/{id}/actions/swipe": {
      "post": {
        "operationId": "CommandsController_swipe",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SwipeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AckCommandResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "DTO 검증 실패·미지원 액션",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "토큰 없음·불일치",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "점유자 불일치",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "기기 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "기기 오프라인",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "요청 과다 (rate limit)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "502": {
            "description": "Agent 미연결·기기 명령 실패",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "504": {
            "description": "기기 응답 시간 초과",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "스와이프",
        "tags": [
          "commands"
        ]
      }
    },
    "/devices/{id}/actions/type": {
      "post": {
        "operationId": "CommandsController_type",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TypeTextDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AckCommandResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "DTO 검증 실패·미지원 액션",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "토큰 없음·불일치",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "점유자 불일치",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "기기 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "기기 오프라인",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "요청 과다 (rate limit)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "502": {
            "description": "Agent 미연결·기기 명령 실패",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "504": {
            "description": "기기 응답 시간 초과",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "텍스트 입력",
        "tags": [
          "commands"
        ]
      }
    },
    "/devices/{id}/actions/ui-dump": {
      "post": {
        "operationId": "CommandsController_uiDump",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UiDumpDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UiDumpCommandResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "DTO 검증 실패·미지원 액션",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "토큰 없음·불일치",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "점유자 불일치",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "기기 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "기기 오프라인",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "요청 과다 (rate limit)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "502": {
            "description": "Agent 미연결·기기 명령 실패",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "504": {
            "description": "기기 응답 시간 초과",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "UI 트리 덤프",
        "tags": [
          "commands"
        ]
      }
    },
    "/devices/{id}/actions/press": {
      "post": {
        "operationId": "CommandsController_press",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PressButtonDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AckCommandResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "DTO 검증 실패·미지원 액션",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "토큰 없음·불일치",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "점유자 불일치",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "기기 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "기기 오프라인",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "요청 과다 (rate limit)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "502": {
            "description": "Agent 미연결·기기 명령 실패",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "504": {
            "description": "기기 응답 시간 초과",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "하드웨어 버튼 (home)",
        "tags": [
          "commands"
        ]
      }
    },
    "/devices/{id}/actions/screenshot": {
      "post": {
        "operationId": "CommandsController_screenshot",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UiDumpDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScreenshotCommandResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "DTO 검증 실패·미지원 액션",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "토큰 없음·불일치",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "점유자 불일치",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "기기 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "기기 오프라인",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "요청 과다 (rate limit)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "502": {
            "description": "Agent 미연결·기기 명령 실패",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "504": {
            "description": "기기 응답 시간 초과",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "화면 캡처 (JPEG base64 + pt 크기)",
        "tags": [
          "commands"
        ]
      }
    }
  },
  "info": {
    "title": "nebula-clone",
    "description": "디바이스 팜 오케스트레이션 API",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ErrorResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 404
          },
          "message": {
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            ]
          },
          "error": {
            "type": "string",
            "example": "Not Found"
          }
        },
        "required": [
          "statusCode",
          "message"
        ]
      },
...
      "TapDto": {
        "type": "object",
        "properties": {
          "occupantId": {
            "type": "string",
            "description": "점유 시 발급된 점유자 ID"
          },
          "x": {
            "type": "number",
            "description": "x 좌표 (pt)"
          },
          "y": {
            "type": "number",
            "description": "y 좌표 (pt)"
          }
        },
        "required": [
          "occupantId",
          "x",
          "y"
        ]
      },
      "AckResultDto": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean"
          }
        },
        "required": [
          "ok"
        ]
      },
      "AckCommandResponseDto": {
        "type": "object",
        "properties": {
          "result": {
            "$ref": "#/components/schemas/AckResultDto"
          }
        },
        "required": [
          "result"
        ]
      },
      "SwipeDto": {
        "type": "object",
        "properties": {
          "occupantId": {
            "type": "string",
            "description": "점유 시 발급된 점유자 ID"
          },
          "fromX": {
            "type": "number"
          },
          "fromY": {
            "type": "number"
          },
          "toX": {
            "type": "number"
          },
          "toY": {
            "type": "number"
          },
          "durationMs": {
            "type": "number",
            "description": "스와이프 시간 (ms, 기본 300)"
          }
        },
        "required": [
          "occupantId",
          "fromX",
          "fromY",
          "toX",
          "toY"
        ]
      },
      "TypeTextDto": {
        "type": "object",
        "properties": {
          "occupantId": {
            "type": "string",
            "description": "점유 시 발급된 점유자 ID"
          },
          "text": {
            "type": "string",
            "description": "입력할 텍스트 (한글·이모지 지원은 Controller 구현에 따름)"
          }
        },
        "required": [
          "occupantId",
          "text"
        ]
      },
      "UiDumpDto": {
        "type": "object",
        "properties": {
          "occupantId": {
            "type": "string",
            "description": "점유 시 발급된 점유자 ID"
          }
        },
        "required": [
          "occupantId"
        ]
      },
      "UiDumpResultDto": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean"
          },
          "tree": {
            "type": "string",
            "description": "접근성 트리 텍스트"
          }
        },
        "required": [
          "ok",
          "tree"
        ]
      },
      "UiDumpCommandResponseDto": {
        "type": "object",
        "properties": {
          "result": {
            "$ref": "#/components/schemas/UiDumpResultDto"
          }
        },
        "required": [
          "result"
        ]
      },
      "PressButtonDto": {
        "type": "object",
        "properties": {
          "occupantId": {
            "type": "string",
            "description": "점유 시 발급된 점유자 ID"
          },
          "button": {
            "type": "string",
            "enum": [
              "home"
            ],
            "description": "하드웨어 버튼 (현재 home만)"
          }
        },
        "required": [
          "occupantId",
          "button"
        ]
      },
      "ScreenshotResultDto": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean"
          },
          "jpegBase64": {
            "type": "string",
            "description": "JPEG 이미지 (base64)"
          },
          "widthPt": {
            "type": "number",
            "description": "화면 폭 (pt) — 탭 좌표 환산 기준"
          },
          "heightPt": {
            "type": "number",
            "description": "화면 높이 (pt)"
          }
        },
        "required": [
          "ok",
          "jpegBase64",
          "widthPt",
          "heightPt"
        ]
      },
      "ScreenshotCommandResponseDto": {
        "type": "object",
        "properties": {
          "result": {
            "$ref": "#/components/schemas/ScreenshotResultDto"
          }
        },
        "required": [
          "result"
        ]
      }
    }
  }
}